DOMEvents/tests/submissions/TestTWF_Paris/DavidBruant/throwing-in-listener-and-window-error-event.html

<!DOCTYPE html>
<!-- Submitted from TestTWF Paris -->
<meta charset="UTF-8">
<title>Throwing in event listener generates an error event on the window object</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<button type="button" id="test">Start test</button>

<script>
    var t = async_test("Throwing in event listener generates an error event on the window object");

    var errorEvent = false;

    window.addEventListener('error', function(e){
        errorEvent = e;
    });

    document.addEventListener('DOMContentLoaded', function(){
        var element = document.getElementById('bim');

        element.addEventListener('click', function(){
            throw new Error('Error from listener');
        });

        element.click();

        assert_true(typeof errorEvent === 'object');
        t.done();
    });
</script>

<div id="bim">
</div>
